requirements.txt
telethon
python-telegram-bot
pillow
python-dotenv